Rename Constants To ReactTodosConstants
Rename Constants to ReactTodosConstants in order not to interfer with
Reacts own Constants file.
